feat: scope caching
